requirements.txt
torch
torchvision
transformers
datasets
evaluate
scikit-learn
matplotlib
pillow
accelerate
huggingface_hub
gradio